pyproject.toml
# SPDX-License-Identifier: Apache-2.0
#
# Licensed to the Apache Software Foundation (ASF) under one
# or more contributor license agreements.  See the NOTICE file
# distributed with this work for additional information
# regarding copyright ownership.  The ASF licenses this file
# to you under the Apache License, Version 2.0 (the
# "License"); you may not use this file except in compliance
# with the License.  You may obtain a copy of the License at
#
#   http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing,
# software distributed under the License is distributed on an
# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
# KIND, either express or implied.  See the License for the
# specific language governing permissions and limitations
# under the License.

[build-system]
requires = ["setuptools >= 61.2"]
build-backend = 'setuptools.build_meta'

[project]
name = "ntfc"
version = "0.0.1"
description = "NTFC for NuttX"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
         "click>=8.1",
         "pyyaml",
         "pyelftools",
         "pexpect",
         "psutil",
         "pyserial",
         "pytest",
         "pytest-dependency",
         "pytest-repeat",
         "pytest-timeout",
         "pytest-ordering",
         "requests",
         "pillow", # optional for graphic?
         "prettytable",
         "pytz", # optional for time zone tests
]
classifiers = [
  "Topic :: Software Development :: Embedded Systems",
  "Operating System :: OS Independent",
]
keywords = [
  "ntfc",
]

[project.scripts]
ntfc = "ntfc.cli.main:main"

[tool.black]
line-length = 79
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 79